+++ src/services/document_service.py
@@ -14,9 +14,10 @@ from common.exceptions import (
 )
 from common.utils.build_complex_search import build_mongo_query
 from common.utils.delete_documents import delete_document
-from common.utils.get_blueprint import get_blueprint_provider, storage_recipe_provider
+from common.utils.get_blueprint import get_blueprint_provider
 from common.utils.get_document_by_path import get_document_uid_by_path
 from common.utils.get_resolved_document_by_id import get_complete_sys_document
+from common.utils.get_storage_recipe import storage_recipe_provider
 from common.utils.logging import logger
 from common.utils.sort_entities_by_attribute import sort_dtos_by_attribute
 from common.utils.string_helpers import split_dmss_ref, split_dotted_id
